feat(cli): new command always pulls latest deps (#381)

+++ tools/test.sh
@@ -1,7 +1,10 @@
 set -e 
 
-NODE_ENV=test PGUSER=postgres ./bin/warthog db:drop
-NODE_ENV=test PGUSER=postgres ./bin/warthog db:create
+if [ -z "$SKIP_DB_CREATION" ]
+then
+  NODE_ENV=test PGUSER=postgres ./bin/warthog db:drop
+  NODE_ENV=test PGUSER=postgres ./bin/warthog db:create
+fi
 
 # Codegen for test files
 NODE_ENV=test ./src/test/codegen-test-files.sh
